+++ src/OutlookGoogleCalendarSync/SyncDirection.cs
@@ -15,8 +15,6 @@ namespace OutlookGoogleCalendarSync {
         public static readonly SyncDirection GoogleToOutlook = new SyncDirection(2, "Outlook ← Google");
         public static readonly SyncDirection Bidirectional = new SyncDirection(3,   "Outlook ↔ Google");
 
-        private SyncDirection() { }
-
         private SyncDirection(int id, String name) {
             this.Name = name;
             this.Id = id;